Switched metadata query parallelization to be gated on CPU count.
Instead of disabling fstatat parallelization based on environment
variables, we instead disable it based on logical CPU count. For
single-core systems, this parallelization infrastructure only adds
overhead. It does immediately see benefits with dual cores, however, so
we enable it for all multicore systems. This may require some tuning as
time goes on.