Inlined full original licenses for consistency.
This commit modifies the two assembly files in the filesystem package
that we've modified/vendored to include their full original licenses.
These already had their original license headers and their full original
licenses were included in Mutagen's license output, but to keep
consistency and encapsulation of this information for vendored files,
we've added the original full license text to the files themselves.