Removed parallel fstatat implementation.
This commit removes the experimental parallel fstatat implementation.
While this implementation showed promise, I think that its risks
outweigh its benefits.

First and foremost, fstatat isn't guaranteed by the POSIX standard to
work with concurrent calls using the same directory file descriptor, so
any "guarantees" of correctness would be based on observation.
Unfortunately it's just not possible to be confident about this, even on
a platform like Linux, simply due to the change in implementations
between kernel versions. There also isn't a good mechanism for rolling
this out on a trial or opt-in basis, and the risks are simply too great.

Second, the implementation itself was really complicated, and it would
almost certainly be broken by changes at some point in the future.
Properly testing it for correctness would have been almost impossible.

Third, the use of a single global Goroutine pool leaves too many open
questions about contention between sessions. I don't think this would
have been an issue, but it could have led to hard-to-debug slowdowns.

Finally, while the performance gains were appreciable, with ~33%
reduction in scan time on multi-core POSIX systems, they also weren't
well-understood across a large phase space of devices, including those
with different core counts and disk speeds. Also, these gains were less
than the loss when moving to the new libSystem-based system calls on
macOS in Go 1.11 and 1.12, which indicates to me that we should wait and
see if the Go runtime is able to optimize these system calls and recoup
some of this performance in Go 1.13 and later.

In any case, this code will live on in the repository history, so it can
be readily recovered if necessary.