Added workaround for aggressive filesystem caching in osxfs.
This commit adds a workaround for the aggressive filesystem caching seen
in issue #73. In particular, it adds the ability to re-open directories
on POSIX systems and takes advantage of this on Linux systems when
attempting to check Unicode decomposition behavior. By re-opening
directories after temporary file creation, new filesystem contents can
be detected.

I also tried to implement a fix by simply forcing a seek-to-beginning
operation on the directory file descriptor, but that also failed to
trigger cache invalidation.