Removed probe modes.
This commit removes the "probe mode" concept from Mutagen. Instead,
we'll continue to rely on automatic probing behavior, now optimized with
fast paths for majority use cases and workarounds for problematic
filesystems.

Worth noting in this removal is that printEndpoint hadn't been updated
to include probe modes, so that should be fixed if reverting this
commit.