Refactored fast path behavior probing and extended to Linux.
This commit refactors how fast path behavior probing mechanisms are
implemented and organized, paving the way for other format-based fast
paths as well as non-format-based fast paths.

This commit also adds fast path support for an initial (and
conservative) set of Linux filesystems.

The structure of code in this commit is not intended to be final, but
additional refactoring (and division) of the filesystem package is
necessary in order to avoid circular dependencies. However, before
performing this refactoring, we need to revert the removal of probe
modes, and reorganizing this code will break that automated reversion.