Relocated filesystem behavior probing and watching code.
This commit relocates the behavior probing and watching code in the
filesystem package to dedicated subdirectories. This is part of an
effort to chop up the filesystem package. Unfortunately, a lot of the
filesystem package's internals rely on low-level primitives that need to
live in the same package, but the behavior probing and watching code
operate at a higher level and are ideal candidates for extraction.

A lot of code has been updated to account for these relocations, but
there shouldn't be much in the way of behavioral changes.