Changed name of QueryFormatByDirectory and deleted TODO comment.
This commit renames QueryFormatByDirectory to QueryFormat (which is
actually the name it already had on platforms which don't use this
function). I think we're probably going to be restricted to the
directory-based/path-based duality for the foreseeable future, so I
don't see any reason to complicate names. Any "ByFile" variants would
have to rely on fstatfs-like mechanisms to work, and I don't see that
being generally useful. Even in the case of file roots, we're best off
just querying behavior for the parent directory.