Relocated filesystem format detection into a subpackage.
This commit further reduces the code volume in the filesystem package by
relocating filesystem format detection into a dedicated filesystem
subpackage. This necessitated adding a Descriptor method to Directory on
POSIX systems in order to access the underlying file descriptor, but
this may be generally useful, so I don't view it as a hack. In order to
keep parity on Windows systems, I've added a corresponding Handle method
to Directory. I've intentionally named and typed these methods
differently in order to ensure they're only used in platform-specific
code.

Since the format detection code is limited in scope, it wasn't generally
useful, and didn't make sense in the filesystem package. It may make
sense to encapsulate it further into an internal package, until such
time that it does become more generally useful (if ever).