Switched away from os.O_* flags in favor of unix.O_* flags.
Since we're using a more raw interface to the filesystem via
golang.org/x/sys/unix, it makes sense to use O_* flags from that package
instead of mixing them with the flags in the os package. At the end of
the day, they have the same definitions, but I prefer a cleaner
separation.