Restructured and expanded handling of TOML-based configuration files.
This commit restructures how TOML-based configuration files are
processed. These files are now processed by the create command and
merged before being sent to the daemon.

This refactoring allows two new features: the ability to ignore the
global configuration file and the ability to specify a custom
configuration file. The merging infrastructure behind this is such that
we could allow an arbitrary number of configuration files to be
specified on the command line, but we'll start with a single file for
the time being. Allowing multiple files to be specified can be done in a
backwards-compatible manner.