Added "neighboring" staging mode and consolidated housekeeping.
This commit adds a "staging mode" concept to Mutagen and establishes two
staging modes: "mutagen" and "neighboring". The former is the standard
staging within the Mutagen data directory, while the latter creates
staging directories that neighbor their corresponding synchronization
roots. This is designed to solve the chronic performance problems that
users experience when the Mutagen data directory resides on a different
filesystem than the synchronization root (and atomic rename operations
have to be emulated).

This commit also consolidates housekeeping functions into a single
package, though this unfortunately brings in a large number of secondary
changes.