Added fast path for Scan when changes are known.
This commit adds a fast path for the scanning infrastructure. This path
processes only the minimal portion of the filesystem necessary to
account for a list of known changes. This requires a baseline snapshot
and a list of modified paths.

This infrastructure will be coupled with watching infrastructure to
allow for ultra-fast rescans.