Added cache backfills for fast scans.
Since fast, baseline-based scans generally won't explicitly traverse
some of the synchronization root, we have to manually populate new
caches to account for unvisited filesystem entries.

For ignore caches, we can't easily know exactly what to propagate from
the old cache, and even if we could, the old cache lookups are
expensive. To work around this issue, we use a best-effort population
mechanism for the new cache where we add false entries for all content
that we see in the result.

For digest caches, we can know exactly which entries need to be
propagated, so we perform that propagation and watch for missing
entries.