Redistributed staging package into session and session/endpoint/local.
The staging package being separate from session and endpoint was
primarily a crutch to work around the session package's dependency on
the configuration package (we needed a third package where the staging
mode could live), but since that dependency no longer exists, it doesn't
make sense for staging to stand on its own as a separate package.

Instead, we've redistributed StagingMode infrastructure into the session
package (and renamed it to StageMode for consistency with other modes)
and moved the stager infrastructure into the session/endpoint/local
package.