Changed default entry count and staging file size limits.
This commit replaces the "zero means unlimited" behavior for the
MaximumEntryCount and MaximumStagingFileSize session parameters.
Instead, the default value in these cases is now the maximum value of a
uint64 type. This was done to fix a theoretical inconsistency with
configuration, where one was unable to recoup the default behavior (of
unlimited counts/sizes) with an explicit specification at a higher level
of configuration. By having a default value that can be explicitly
specified by users, the default behavior can be recovered. This works
without breaking existing sessions, and the new default values are so
large that they are practically unlimited on modern computers.