Cleaned up printing of large default maximums.
Since we'll likely be using 2^64 - 1 as a default maximum value for a
number of parameters, this commit factors the value out into its own
package (again, to avoid the dependency on the large math package) and
adds a nice human-readable description. We may add more numeric
constants in the future. For the time being, these constants have been
used to make session printing a little bit more human-friendly.