Added check for missing stager files in Transition.
This commit modifies Transition to watch for missing files when calling
Provider.Provide. This is almost certainly an indication of concurrent
modifications during staging. This requires cooperation from the
Provider implementation and is documented for the Provider.Provide
method. Fortunately this is easy to implement for stager.

If missing files are detected, controller will now force another
synchronization cycle immediately, with the idea being that the
concurrent modifications are ephemeral and another round of staging
should solve the problem.