Modified controller to track reason for skipping polling.
This commit modifies the controller's synchronization loop to track the
reason for skipping polling, if any. This allows us to do two important
things: First, it allows us to skip the rescan wait period on at least
the first scan retry, which should make for a significantly better user
experience. Second, it allows us to avoid skipping polling more than
once in the case of missing stager files - if the files continue to be
missing, users can wait for the filesystem to stabilize and then force
synchronization manually with the flush command.