Revert "Disabled GO111MODULE environment variable on CI systems."
This reverts commit d5b854002cb225f38de996049aa3745dccf25c6b.

It turns out that this variable is somewhat necessary, even with Go
1.12, because if a CI system clones Mutagen into $GOPATH/src (which
happens on Travis), and GO111MODULE is unset or set to "auto" (which is
the default and also explicitly done on Travis), then module support is
disabled. Until GO111MODULE=on is the default, let's continue to set
this, even on AppVeyor (where it may not be necessary since we don't
clone into %GOPATH%/src).

Looking forward, it looks like Go 1.13 may not make GO111MODULE=on the
default after all, but it looks like it may remove the
disable-if-in-$GOPATH/src behavior, so we may be able to get rid of this
setting at that point. See golang/go#31857 for more information.