Changed root command description.
Mutagen's functionality is going to be extended in the near future, so
this commit changes the help description of the root command to better
reflect Mutagen's ambitions. It also removes the full stop from the root
command description for consistency with the rest of the command
descriptions and updates the punctuation of the mutagen-agent root
command description to match.