Relocated and refactored Docker and SSH invocation mechanisms.
This commit extracts the Docker and SSH invocation mechanisms from their
respective protocol implementations and refactors/exposes this code in a
separate tools package. This will allow non-protocol code to use these
tools, and the addition of a context argument to invocation should allow
Mutagen to better control subprocess lifecycles.