Aligned POSIX and Windows IPC behavior, added more explicit permissions.
This commit attempts to align POSIX and Windows IPC behavior by
requiring that a pipe name record file doesn't already exist on Windows
and doing away with stale socket removal on POSIX. This allows listening
on an IPC endpoint to act as a lock of sorts at the cost of having to
manually remove stale sockets in the case of a program crash. In the
context of the Mutagen daemon (where a separate lock file exists),
automatic file removal can still occur externally, but this shouldn't be
the default behavior of the IPC package.

This commit also adds explicit permission bits for sockets on POSIX
systems. These were previously computed by umask, which generally
disables write bits (and thus connectivity) for other users, and the
socket was also restricted by its prefix path having search permissions
disabled for other users' processes (also required for connectivity),
but we may start using the IPC infrastructure outside of the daemon
directory, and thus need more explicit permissions.