Simplified stale IPC endpoint removal for daemons.
As of Go 1.12, os.IsExist doesn't seem to detect cases where listening
on a socket fails due to the socket already existing. This is probably
fixed in Go 1.13 with the new error unwrapping functionality, but in the
mean time (and probably indefinitely) we can just do a preemptive
removal of the endpoint.