Decoupled agent transports from session protocols.
We're going to want to use the agent mechanism for protocols other than
the remote endpoint protocol, so this commit decouples the protocol and
transport layers so that transports reside within the agent package and
protocols reside within the session package. Thus, the agent package no
longer knows anything about the session package, and it can be re-used
to support other protocols.