Removed incorrect error checking in agent dialing.
It turns out that we don't want to check for an error from
process.Connection.Close, because the error from the exit of the
underlying process is returned (not the error from terminating the
process). This commit removes the incorrect error check that we added
when decoupling protocols from transports.