Added the ability to create ephemeral synchronization sessions.
This commit modifies the session package to enable to creation of
ephemeral sessions. Ephemeral sessions avoid writing any files to disk
(with the exception of endpoint staging directories, which are forced
into neighboring mode) and thus terminate when the corresponding
Controller is shut down.

This commit also starts threading logging facilities into the session
package.