Changed how sync commands are registered.
In order for the sync commands to have the correct help message, they
need to have the sync command as their parent, but the way that we were
registering them with the root command for backwards compatibility was
reparenting the commands onto the root commands. This commit modifies
the way that these commands are registered with the root command so that
they have the correct parent.