Renamed synchronization path constants.
This commit renames the synchronization path constants to reflect the
fact that we're going to create other session types. Unfortunately we
can't change their values without breaking compatibility, so their
existing names will live on as a historical artifact. Ideally we
probably would have structured these directories hierarchically under a
"synchronization" directory. We should change this if we have a
compatibility-breaking release.