Disabled gRPC support in protoc-gen-go when not necessary.
We only need gRPC support for our service packages, so in order to avoid
inadvertently pulling in a gRPC dependency into our other packages
(where it would bloat agent binaries), we disable support for gRPC
outside of the service packages. This doesn't affect any generated code,
because the protoc-gen-go generator is smart enough to know not to
include a gRPC dependency needlessly, but this should explicitly guard
against any inclusion.