Continued threading logging into synchronization.
This commit continues threading logging into the synchronization manager
and controllers. The logging is basic and minimal for now, but the
infrastructure is present to vastly expand the logging (as necessary).

Making this change necessitated computing session identifiers in the
manager and passing them to the session controller, but this is a better
design anyway.