Renamed forwarding command to forward.
While forwarding might be better because it was a noun on which verb
subcommands could operate, it's just too long to type, and consistency
would warrant renaming the sync command to synchronization. This change
lets us keep these two commands as verbs that might potentially operate
in their own right to create ephemeral sessions.