Reassigned responsibility for version handshake to agent.
This commit relocates the version handshake to take place as part of the
agent dialing process. This is a more appropriate location and means
that the version handshake also takes place for forwarding endpoints. We
purposefully keep the version handshake separate from the agent
handshake so that we can classify those issues separately.