Added integration tests for forwarding.
This commit adds a minimal integration test for forwarding sessions. It
creates a simple HTTP server, which fortuitously serves as a better
Docker container entrypoint, and tests that this server can be accessed
via a forwarding session into a Docker container.

In order to support this new test, this commit also adds a forwarding
session manager to the integration test daemon. It also implements a
netpipe protocol for forwarding, though this isn't currently used.