Added flag that switches Mutagen data directory for development
This commit adds a flag (mutagen.DevelopmentVersion) that encodes
whether or not the current build represents a development version of
Mutagen. If this flag is set to true, then the Mutagen data directory
will be set to ~/.mutagen-dev for the daemon, agents, and test package
builds.

The goal of this commit is to make it easier to use Mutagen to develop
itself.

Signed-off-by: Jacob Howard <jacob@havoc.io>