Reinstated TOML parsing in a legacy capacity and restructured YAML
It's a big ask to force our users to immediately switch to YAML, so this
commit adds support for falling back to legacy TOML parsing
automatically (with a warning printed). We're not going to extend the
TOML configuration going forward, and we still only support it for
synchronization sessions (since we never even defined a TOML format for
forwarding sessions), but we're not going to phase it out in a single
version.

Looking towards future extensibility, we've relocated the forwarding and
synchronization session defaults under appropriate configuration keys in
YAML configuration files. This will allow us to define full sessions in
YAML later.

Signed-off-by: Jacob Howard <jacob@havoc.io>