Made session path normalization more consistent
This commit modifies the synchronization and forwarding services to
enforce that local paths (for synchronization roots and socket paths)
are normalized when they arrive.

It also modifies forwarding endpoint servers to perform socket path
normalization.

Finally, it relocates synchronization root path normalization from the
local endpoint constructor to the remote endpoint server, which is the
only place that it should be necessary given that local paths are
pre-normalized.

Signed-off-by: Jacob Howard <jacob@havoc.io>