Removed comment about reason for using separate initialization stream
Our remote forwarding protocol uses a separate multiplexer stream for
initialization instead of initializing directly on the underlying
connection. While this is a good sanity check for multiplexing, the
reason for doing this that was stated in the comments (about our
Protocol Buffers decoder using a buffered reader) didn't make sense. A
buffered reader would still only be able to read until the end of any
transmitted messages, so it would still leave the line in a predictable
state. In any case, we'll leave this implementation for now, but not the
incorrect comment.

Signed-off-by: Jacob Howard <jacob@havoc.io>