Modified version handshake to only require match to minor version.
Since we follow semantic versioning, we really only need to match agent
binaries on minor version. We'll continue looking for exact matches when
using agent-based transports (since we invoke a path that contains the
full version), but for tunnel host images, we don't want to have to
carry around every possible patch version of a particular release, so we
allow a little bit of version skew.

Signed-off-by: Jacob Howard <jacob@havoc.io>