Added single-message Protocol Buffers encoding/decoding routines.
Since our Protocol Buffers decoder implementation uses a buffered reader
around the underlying stream, we don't have any way to force it to read
only to message bounaries. This is problematic if we want to start a
stream with a Protocol Buffers message but then switch to a different
protocol. This commit adds a single-message decoding function that can
avoid over-reading a stream at the cost of a small amount of
performance. For single-message initialization, it's a perfect fit. To
complement this function, we've added a corresponding single-message
encode function, though this is merely a convenience function designed
to keep symmetry.

Signed-off-by: Jacob Howard <jacob@havoc.io>