Refactored reconnection waits in synchronization and forwarding.
This commit reduces the reconnection wait time in synchronization and
forwarding to 15 seconds. More importantly, it also adds instant
redialing on synchronization/forwarding failure, so long as the last
failure was longer ago than the reconnection wait time. This allows for
better utilization of tunnels' blocking dialing mechanism without
introducing unchecked CPU usage in the case of persistent errors.

Signed-off-by: Jacob Howard <jacob@havoc.io>