Disabled user specification in tunnel URLs.
User specification was never supported in tunnels, so there was no point
in having the code in place to handle it. Having tunnel agents run as
the hosting user is the better security model in any case. If we do
decide to add this back, we'll need to create a new tunnel version
anyway.

Signed-off-by: Jacob Howard <jacob@havoc.io>