Refactored the manner in which project lock files are removed.
It turns out that Windows won't allow lock file removal, even by the
same process, if the file is locked or even just open. We can invoke
os.Remove after closing the file, but this leads to a race condition and
also doesn't work on POSIX systems (where locks are merely advisory and
files can be unlinked even if open by other processes). Consequently, we
have to handle this removal in a deferred, platform-specific fashion,
and we have to use a few hacks to avoid race conditions on Windows.
Fortunately Windows doesn't enable FILE_SHARE_DELETE by default and
its locks are not merely advisory, so as long as we truncate lock files
before releasing them, there is no race condition in this case.

We've also added more aggressive removal of lock files when project
initialization fails before any resources are actually created.

Signed-off-by: Jacob Howard <jacob@havoc.io>