Renamed and relocated sidecar container provisioning script.
This provisioning script can actually serve containers that aren't using
tunnels. For example, we could make the Docker transport attempt to dial
a Mutagen agent in /usr/libexec. This will require more development and
thinking, and we'll probably rewrite the provisioning script as a
portable Go binary at some point, but for the time being I think the
script should be stored in a more generic location.

Signed-off-by: Jacob Howard <jacob@havoc.io>