Fixed colorized usage and error output on Windows.
On Windows, the github.com/fatih/color package requires a special output
device for escape sequences to work. We were using this (either directly
or indirectly) for colorized console output, but we weren't telling
Cobra to use this for its usage and error output. This commit enables
colorized output for all Cobra commands in the CLI command hierarchy.

Fixes #153.

Signed-off-by: Jacob Howard <jacob@havoc.io>