Switched conflict printing to use names instead of Greek letters.
The α and β characters used for printing conflicts are concise and cute,
but they can cause problems in environments where Unicode isn't fully
supported, and they run the risk of not being widely understood. This
commit switches to using full endpoint names instead of these
characters. Some padding is necessary due to the differing endpoint name
lengths, but conflicts should be easier to read now.

Fixes #154.

Signed-off-by: Jacob Howard <jacob@havoc.io>