Implemented "reset" command for synchronization sessions.
This commit adds a "reset" command for synchronization sessions that
allows users to clear the ancestor data for a synchronization session.
This is useful in cases where storage on one endpoint has been reset
(e.g. when a container volume that was recreated) and the user wants to
once again rely upon the initial both-created-same behavior without
having to recreate the session (or project). In the wake of our new
safety checks, this will probably be a useful convenience.

Signed-off-by: Jacob Howard <jacob@havoc.io>