Added additional integration test fixes for asynchronous connections.
This commit adds additional integration test fixes for the asynchronous
connection behavior adopted in b44f284.

First, we've disabled concurrent test execution for Docker-related tests
on Windows. This is necessary because Docker agent installation on
Windows requires halting the container, meaning that two Docker-related
tests could conflict with each other. This was actually required
regardless of the changes in b44f284.

Second, we've marked the Docker protocol as potentially requiring user
input, because for Windows containers it can require confirmation that
the user wants to temporarily halt the container to install an agent
binary.

Signed-off-by: Jacob Howard <jacob@havoc.io>