Added additional test fix for b44f284 changes.
This commit adds yet another integration test fix to account for the
changes in b44f284. We now create the forwarding session in
TestForwardingToHTTPDemo in a paused state. This works around a sequence
of events that's triggered by the race detector that, while still
behaving correctly, causes the test (as written) to fail.

Signed-off-by: Jacob Howard <jacob@havoc.io>