Removed uid/gid lookup check in NewOwnershipSpecification.
When dealing with containers, it may be the case that ownership IDs are
set for users and groups that don't exist in one container but do exist
in another. As such, we won't always be able to look up the
corresponding user/group from the ID. To allow Mutagen to work with
these and related cases, this commit removes the LookupId and
LookupGroupId calls that were previously used to validate raw IDs on
POSIX systems.

Signed-off-by: Jacob Howard <jacob@havoc.io>