Implemented write closure EOF transmission for network forwarding.
We weren't previously transmitting EOFs arising from write closure on
connections (instead only sending those on full connection closure). We
were also prematurely closing connections when one side performed a
write-only closure. This commit fixes that problem with a smarter
forwarding loop that monitors for write closure. In order to facilitate
this, we had to add a wrapper to enable write closure for multiplexer
streams (since they don't support write closure natively). This comes at
the cost of some performance, though it appears to be negligible. We
might be able to avoid this wrapper in the future by adding CloseWrite
support to yamux streams.

Signed-off-by: Jacob Howard <jacob@havoc.io>