Added explicit incoming connection closure on forwarding termination.
This commit adds an additional incoming connection closure in the case
of multiplexer failure on a remote forwarding endpoint. This isn't that
important, because we generally exit the agent process in this case
anyway, but technically the forwarding could be embedded into a larger
process, and while the ForwardAndClose Goroutines would close existing
connections, the new incoming connection would just be left open.

Signed-off-by: Jacob Howard <jacob@havoc.io>