Refactored identifier creation to ensure consistent length.
This commit refactors the way that identifiers are created in order to
ensure a consistent total length of 48 characters without sacrificing
entropy. The core change is that the Base62-encoded portion of the
identifier is now left-padded with 0s to ensure it reaches the target
length. These 0s have no effect on the underlying encoded value.

Signed-off-by: Jacob Howard <jacob@havoc.io>