Added path information to scan errors.
One of the biggest problems we've had with debugging scan issues is
figuring out the path causing the issue, because, unlike Transition,
Scan is pass/fail and doesn't report problematic paths. With our
previous error handling framework, it was tedious to embed this
additional information (though we probably still should have done it),
but with Go 1.13 it's trivial.

Signed-off-by: Jacob Howard <jacob@havoc.io>