Added path information to scan errors.
One of the biggest problems we've had with debugging scan issues is
figuring out the path causing the issue, because, unlike Transition,
Scan is pass/fail and doesn't report problematic paths. This commit
backports the error path information added in v0.11.x (updated for the
older pre-Go 1.13 error framework we're using in v0.10.x).

Signed-off-by: Jacob Howard <jacob@havoc.io>