Added cache for filesystem behaviors determined via probe files.
This commit adds a cache for scans to store filesystem behavior
information determined using probe files. This cache is necessary to
avoid watching/scanning feedback loops caused by probe files. It isn't
used in cases where probe files aren't used. A fix for these feedback
loops was previously attempted in c4830be, but this didn't take into
account the modification time changes that are seen on probe files'
parent directories. Instead, we now don't attempt to ignore probe files
when watching, but rather rely on the behavior cache to simply avoid
creating them a second time, thereby averting a feedback loop.

This design does come with the remote possibility that a device ID could
be re-used with a different filesystem with different behavior, but
that's quite unlikely given POSIX's requirements for device IDs. Caching
is also only used in select and uncommon cases (since probe files are
rarely used), so the cross-section for this scenario is quite small.

There's also the issue that this cache is currently unbounded in size,
but this shouldn't be an issue in practice given the finite number of
filesystems on a system.

Updates #189, #108, and #44.

Signed-off-by: Jacob Howard <jacob@havoc.io>