Added hooks for setup and teardown commands in projects.
To make it easier to couple session and infrastructure lifecycles, this
commit adds support for setup and teardown commands in Mutagen projects.
The support is fairly basic, but it should be useful for the vast
majority of cases. More complex logic can be implemented in the
specified commands themselves (e.g. using shell scripts).

Signed-off-by: Jacob Howard <jacob@havoc.io>