Improved transport failure monitoring for forwarding.
Forwarding sessions, unlike synchronization sessions, don't have a
simultaneous polling operation on each endpoint. Instead, they poll for
incoming connections from the source and then open outgoing connections
to the destination, meaning that they may not be aware of destination
transport failures until after accepting an incoming connection and
trying to perform a corresponding dial.

To improve upon this, this commit extends the forwarding endpoint
interface with a transport error notification channel and then updates
the forwarding run loop to monitor this channel for each endpoint.
Technically this is only necessary for the destination endpoint, since
the source endpoint is being monitored by waiting for an inbound
connection, but the symmetric monitoring is better and will likely lead
to more direct error reporting since the error will be directly from the
underlying connection rather than from the multiplexer (though which is
reported first is still probabilistic).

Signed-off-by: Jacob Howard <jacob@havoc.io>