Updated polling to improve transport failure detection in no-watch mode.
This commit updates the synchronization loop polling mechanism to always
perform polling, even in the case of endpoints with a no-watch watch
mode. We still ignore events from these endpoints (which shouldn't be
sent in the well-behaved case anyway), but by polling on these
endpoints, we can better detect transport failures while the session is
sitting idle.

Signed-off-by: Jacob Howard <jacob@havoc.io>