Cleaned up contexts used in tunnel dialing handoff.
The tunnel controller's dial method previously created an additional
subcontext to ensure that dialing was cancelled when the function
returned. The idea behind this was to ensure that future refactors
couldn't block the serving loop, which relies on the dial function not
returning after the submission of a dialing request except in cases of
cancellation. This extra subcontext wasn't necessary in the current
implementation, and the associated comment was sort of confusing. This
commit just restructures the code to make the coupling of these two
operations more apparent, and adds a clearer comment about the need for
this behavioral invariant (along with suggestions about the subcontext
approach in case it's needed in the future).

Signed-off-by: Jacob Howard <jacob@havoc.io>