Made synchronization controller's use of contexts more idiomatic.
Usage of contexts was sort of shoehorned in to the synchronization
controller since they came later. As such, the controller's usage of
contexts wasn't particularly idiomatic, both in terms of naming and
positioning in function arguments. This commit takes the opportunity to
correct that.

Additionally, since there's no longer a name collision with the sync
package, we've removed the previous alias.

Signed-off-by: Jacob Howard <jacob@havoc.io>