Made all connection operations preemptable.
This commit pushes down preemption from the connect layer in forwarding
and synchronization to the protocol handler layer. It also unifies all
connect operations to use the same method while supporting preemption.
In the case of tunnel protocol handlers, this was a simple addition,
since tunnel dialing operations already support preemption. In the case
of agent-based protocol handlers, we've just relocated the mechanism
that we used in the connect layer (a background Goroutine that simply
closes its connection on cancellation) into the protocol handlers
themselves. This results in some code duplication (about a factor of
two), but it saves us the complexity of having to implement this
preemption in the agent package, where it's more difficult to find a
"good" point to preempt. With the strategy we've chosen, agent
installation operations may continue in the background, but this is more
desirable than leaving agent binaries scattered about if cancellation
occurs during an scp or docker cp operation (which is exactly where it's
most likely to occur).

Signed-off-by: Jacob Howard <jacob@havoc.io>