Restructured remote synchronization endpoint client Poll method.
This commit restructures how context cancellation is invoked in the
remote synchronization endpoint client's Poll method. The old method was
correct, but it used a defer on one path and an explicit invocation on
another path and it was confusing (to the point where it looked like
there was a race condition even though there wasn't).

Signed-off-by: Jacob Howard <jacob@havoc.io>